pyproject.toml
[tool.poetry]
name = "cleanform"
version = "0.1.0"
description = ""
authors = ["Santosh Kumar Patro <patrosantoshkumar1212@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"
django-cors-headers = "^3.13.0"
python-dotenv = "^0.20.0"
psycopg2-binary = "^2.9.3"
nanoid = "^2.0.0"
autopep8 = "^1.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
